+++ NEWS.md
@@ -1,9 +1,14 @@
-# ggdistribute 1.0.0.0000
+# ggdistribute 1.0.1.9000
+
+* Fix R CMD Check notes
+* Align geoms to y axis
+* Extend README
+* Fix reverse in position_spread
+
+# ggdistribute 1.0.0
 
 * Upload initial version of package to GitHub
-
 
 # ggdistribute 0.0.0.9001
 
 * Added a `NEWS.md` file to track changes to the package.
-